Quash gcc's warnings with -Wall -Wextra -Wno-unused -Wno-missing-field-initializers. Practically all of these were signedness mismatches, and all look harmless except possibly for the ircd_in_addr_valid() one.
git-svn-id: file:///home/klmitch/undernet-ircu/undernet-ircu-svn/ircu2/trunk@1858 c9e4aea6-c8fd-4c43-8297-357d70d61c8c

+++ ircd/gline.c
@@ -140,7 +140,7 @@ count_realnames(const char *mask)
   for (acptr = GlobalClientList; acptr; acptr = cli_next(acptr)) {
     if (!IsUser(acptr))
       continue;
-    if (strlen(cli_info(acptr)) < minlen)
+    if (strlen(cli_info(acptr)) < (unsigned int)minlen)
       continue;
     if (!matchexec(cli_info(acptr), cmask, minlen))
       count++;
@@ -461,7 +461,7 @@ gline_add(struct Client *cptr, struct Client *sptr, char *userhost,
   } else {
     canon_userhost(userhost, &user, &host, "*");
     if (sizeof(uhmask) <
-	ircd_snprintf(0, uhmask, sizeof(uhmask), "%s@%s", user, host))
+	(unsigned int)ircd_snprintf(0, uhmask, sizeof(uhmask), "%s@%s", user, host))
       return send_reply(sptr, ERR_LONGMASK);
     else if (MyUser(sptr) || (IsUser(sptr) && flags & GLINE_LOCAL)) {
       switch (gline_checkmask(host)) {
